+++ jgiven-tests/src/main/java/com/tngtech/jgiven/tests/PendingDataProviderTests.java
@@ -5,10 +5,6 @@ import com.tngtech.java.junit.dataprovider.DataProviderRunner;
 import com.tngtech.jgiven.annotation.Pending;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.testng.annotations.Listeners;
-
-import com.tngtech.jgiven.annotation.NotImplementedYet;
-import com.tngtech.jgiven.testng.ScenarioTestListener;
 
 @RunWith(DataProviderRunner.class)
 public class PendingDataProviderTests extends ScenarioTestForTesting<GivenTestStage, WhenTestStage, ThenTestStage> {